requirements.txt
chromadb
openai
python-dotenv
pandas